Height and window resize calculation on the item list
