+++ app/assets/javascripts/components.js
@@ -7,6 +7,7 @@ React = require('react');
 // Mixins
 CollectionUrlMixin = require('./mixins/CollectionUrlMixin');
 HorizontalScrollMixin = require('./mixins/HorizontalScrollMixin');
+PageHeightMixin = require('./mixins/PageHeightMixin');
 PrevNextMixin = require('./mixins/PrevNextMixin');
 
 // Layout
